requirements.txt
pydantic==2.11.4
pydantic-settings==2.9.1
google-generativeai==0.8.5
